workspace: Detect git worktrees as colocated workspaces
Previously, only the default workspace could be colocated. This adds
logic to the colocation-detection code to support finding a worktree
with a "gitfile" as well as a full-on .git directory.

There are some new tests for what happens when you `jj git init
--git-repo=...` either in or pointing at an existing worktree. I do not
expect these to be common workflows, but there is new behaviour here
that we need to track.