git: get_git_repo -> get_git_backend_repo, distinguish it from current worktree
Now that colocated workspaces exist, it is necessary to distinguish the
two kinds of git repository `jj` might open. There's the git backend
(which is either a bare repository or a colocated one) and the worktrees
for all the colocated workspaces. Both of these are a
`gix/git2::Repository`, the difference is just which path you opened. So
you have to distinguish them with names.

This rename mostly serves as an opportunity to review at least some of
the usages of explicit usage of the git backend repo.